+++ CHANGELOG.md
@@ -10,3 +10,11 @@ All notable changes to this project will be documented in this file.
 - Stackable and expandable toasts
 - Toast styles
 - Smooth toasts
+
+### Fixed 0.7.2
+
+- Pointer events are applied to hidden overflowing toasts. Fixed it.
+
+### Reduced Bundle Size 0.7.3
+
+- Dynamically imported icons for a smaller initial bundle size.